ofmcc-851 - refactor application view + remove hover behaviour for table headers in Messages and Notifications tables
