+++ frontend/src/views/applications/ApplicationsHistoryView.vue
